cinemaslides: Encrypted DCPs. KDM mode
Option --encrypt will generate encrypted essence trackfiles.
Content keys are generated on the fly and will be stored in
$CINEMASLIDESDIR/keys/. For now the format of keyfiles is

<Key UUID>:<Key type>:<Key data>
16710c91-df17-41a5-9d77-02d491ea933a:MDIK:6cb7686e927bc1540a31797b2e157273

and filenames are the respective key UUIDs. The keyfile format is
somewhat prototype legacy and will quite likely change to key
data only.

Option --kdm, along with --cpl, --target, --start, --end,
--annotation and --issuer, will generate a key delivery
message (KDM) for a given CPL, targeting a given (server)
certificate with a given validity period.

For now it works only with the keyfile format described above.

As of v0.2010.11.29 KDM code is in a kind of, what I'd call it,
blissful show-stopped state: It creates KDMs which ingest
successfully on our XDC Cinestore Solo G3 server which will
then display the correct time window and a nifty 'unlocked' icon.
Select for playback fails, though. Work in progress ...