cinemaslides: KDM options --cpl-id and related
As an alternative to generating KDMs with direct access to the CPL file
this patch allows to specify, along with a CPL ID, a keys directory
which contains the content keys for the CPL in question. In addition use
--cpl-content-title to specify the title of the composition which is,
among other things, used as one of the particles of the KDM filename.

Format of keyfiles required by cinemaslides is unchanged:
  <key id>:<key type>:<key value>
Names of keyfiles required by cinemaslides is unchanged:
  <key id>

Not heavily tested as in it does what it is supposed to do but it won't
catch all and any malformed input from the user's end.