cinemaslides: Check --cpl-id for RFC 4122 compliance
Check the parameter to --cpl-id for RFC 4122 compliance. Until this fix
it was possible to specify a path to a CPL file which would end up in the
generated KDM -- which in turn would render the KDM invalid.

Note: The --cpl-id option is useful only when there is no direct access
to the CPL file. In almost all cases using --cpl <path to cpl file> is
the better option.